requirements.txt
annotated-types==0.6.0
anyio==4.3.0
bcrypt==4.1.2
boto3==1.34.84
botocore==1.34.84
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
decorator==4.4.2
ecdsa==0.18.0
exceptiongroup==1.2.0
fastapi==0.110.0
greenlet==3.0.3
h11==0.14.0
httptools==0.6.1
idna==3.6
imageio==2.34.0
imageio-ffmpeg==0.4.9
jmespath==1.0.1
moviepy==1.0.3
numpy==1.26.4
passlib==1.7.4
pillow==10.3.0
proglog==0.1.10
psycopg2==2.9.9
pyasn1==0.5.1
pydantic==2.6.4
pydantic_core==2.16.3
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
s3transfer==0.10.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.28
starlette==0.36.3
tqdm==4.66.2
typing_extensions==4.10.0
urllib3==2.2.1
uvicorn==0.29.0
watchfiles==0.21.0
websockets==12.0
